Move all mock alert code into a common testutils file
